requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.2.0
numpy>=1.23.5
pandas>=2.0.0
PyYAML>=6.0.0
scipy>=1.10.0
tabulate>=0.9.0
opencv-python>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pycocotools>=2.0.6
tqdm>=4.65.0
Flask>=2.3.0
flask-cors>=4.0.0